vmstat as long
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@20609 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ uvm-lib/impl/com/untangle/uvm/engine/MessageManagerImpl.java
@@ -485,7 +485,7 @@ class MessageManagerImpl implements LocalMessageManager
                         if (keepers.contains(n)) {
                             String s = matcher.group(2);
                             try {
-                                m.put(n, Integer.parseInt(s) * multiple);
+                                m.put(n, Long.parseLong(s) * multiple);
                             } catch (NumberFormatException exn) {
                                 logger.warn("could not add value for: " + n);
                             }
